Move LinearEquation, Matrix, and RHS into a separate module, add documentation
