+++ docs/manual.tex
@@ -1206,7 +1206,7 @@ with constructor arguments
 The value of a matrix action $A x$ is defined by the \texttt{forward\_action}
 method. The \texttt{forward\_action} method has the form
 \begin{lstlisting}
-def forward_action(self, nl_deps, x, b, method="assign"):
+def forward_action(self, nl_deps, x, b, *, method="assign"):
 \end{lstlisting}
 with arguments
 \begin{itemize}
@@ -1232,7 +1232,7 @@ The value of a matrix action $A^* x$ is defined by the
 \texttt{adjoint\_action} method. The \texttt{adjoint\_action} method has the
 form
 \begin{lstlisting}
-def adjoint_action(self, nl_deps, adj_x, b, b_index=0, method="assign"):
+def adjoint_action(self, nl_deps, adj_x, b, b_index=0, *, method="assign"):
 \end{lstlisting}
 with arguments
 \begin{itemize}
@@ -1287,7 +1287,7 @@ The action of the adjoint of the derivative of $A x$ is defined by the
 \texttt{adjoint\_derivative\_action} method. The
 \texttt{adjoint\_derivative\_action} method has the form
 \begin{lstlisting}
-def adjoint_derivative_action(self, nl_deps, nl_dep_index, x, adj_x, b,
+def adjoint_derivative_action(self, nl_deps, nl_dep_index, x, adj_x, b, *,
                               method="assign"):
 \end{lstlisting}
 with arguments
@@ -1330,10 +1330,10 @@ terms), a \texttt{RHS}, or a sequence of \texttt{RHS} objects.
 A \texttt{Matrix} may act on multiple functions. In this case methods have the
 form
 \begin{lstlisting}
-def forward_action(self, nl_deps, X, B, method="assign"):
+def forward_action(self, nl_deps, X, B, *, method="assign"):
 \end{lstlisting}
 \begin{lstlisting}
-def adjoint_action(self, nl_deps, adj_X, b, b_index=0, method="assign"):
+def adjoint_action(self, nl_deps, adj_X, b, b_index=0, *, method="assign"):
 \end{lstlisting}
 \begin{lstlisting}
 def forward_solve(self, X, nl_deps, B):
@@ -1342,7 +1342,7 @@ def forward_solve(self, X, nl_deps, B):
 def adjoint_solve(self, adj_X, nl_deps, B):
 \end{lstlisting}
 \begin{lstlisting}
-def adjoint_derivative_action(self, nl_deps, nl_dep_index, X, adj_X, b,
+def adjoint_derivative_action(self, nl_deps, nl_dep_index, X, adj_X, b, *,
                               method="assign"):
 \end{lstlisting}
 \begin{lstlisting}
