+++ tlm_adjoint/__init__.py
@@ -29,6 +29,7 @@ from .hessian import GeneralGaussNewton as GaussNewton  # noqa: F401
 from .hessian import GeneralHessian as Hessian  # noqa: F401
 from .hessian_optimization import *  # noqa: F401
 from .interface import *             # noqa: F401
+from .linear_equation import *       # noqa: F401
 from .manager import *               # noqa: F401
 from .optimization import *          # noqa: F401
 from .overloaded_float import *      # noqa: F401
