+++ tlm_adjoint/equations.py
@@ -19,19 +19,16 @@
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
 from .interface import check_space_types, check_space_types_conjugate_dual, \
-    check_space_types_dual, conjugate_dual_space_type, function_assign, \
-    function_axpy, function_comm, function_dtype, function_get_values, \
-    function_id, function_inner, function_local_size, \
-    function_new_conjugate_dual, function_replacement, function_set_values, \
-    function_space, function_space_type, function_sum, function_zero, \
-    is_function, space_new
+    check_space_types_dual, function_assign, function_axpy, function_comm, \
+    function_dtype, function_get_values, function_id, function_inner, \
+    function_local_size, function_new_conjugate_dual, function_replacement, \
+    function_set_values, function_sum, function_zero, is_function
 
 from .alias import WeakAlias
-from .equation import Equation, Referrer
+from .equation import Equation, ZeroAssignment
+from .linear_equation import LinearEquation, RHS
 from .tangent_linear import get_tangent_linear
 
-from collections.abc import Sequence
-import inspect
 import numpy as np
 import warnings
 
@@ -40,11 +37,6 @@ __all__ = \
         "Assignment",
         "Axpy",
         "LinearCombination",
-        "ZeroAssignment",
-
-        "LinearEquation",
-        "Matrix",
-        "RHS",
 
         "DotProductRHS",
         "DotProduct",
@@ -60,64 +52,10 @@ __all__ = \
         "MatrixActionSolver",
         "NormSqRHS",
         "NormSqSolver",
-        "NullSolver",
         "ScaleSolver",
         "SumRHS",
         "SumSolver"
     ]
-
-
-class ZeroAssignment(Equation):
-    r"""Represents an assignment
-
-    .. math::
-
-        x = 0.
-
-    The forward residual is defined
-
-    .. math::
-
-        \mathcal{F} \left( x \right) = x.
-
-    :arg X: A function or a :class:`Sequence` of functions defining the forward
-        solution :math:`x`.
-    """
-
-    def __init__(self, X):
-        if is_function(X):
-            X = (X,)
-        super().__init__(X, X, nl_deps=[], ic=False, adj_ic=False)
-
-    def forward_solve(self, X, deps=None):
-        if is_function(X):
-            X = (X,)
-        for x in X:
-            function_zero(x)
-
-    def adjoint_derivative_action(self, nl_deps, dep_index, adj_X):
-        if is_function(adj_X):
-            adj_X = (adj_X,)
-        if dep_index < len(adj_X):
-            return adj_X[dep_index]
-        else:
-            raise IndexError("dep_index out of bounds")
-
-    def adjoint_jacobian_solve(self, adj_X, nl_deps, B):
-        return B
-
-    def tangent_linear(self, M, dM, tlm_map):
-        return ZeroAssignment([tlm_map[x] for x in self.X()])
-
-
-class NullSolver(ZeroAssignment):
-    ""
-
-    def __init__(self, X):
-        warnings.warn("NullSolver is deprecated -- "
-                      "use ZeroAssignment instead",
-                      DeprecationWarning, stacklevel=2)
-        super().__init__(X)
 
 
 class Assignment(Equation):
@@ -293,446 +231,6 @@ class AxpySolver(Axpy):
                       "use Axpy instead",
                       DeprecationWarning, stacklevel=2)
         super().__init__(y_new, y_old, alpha, x)
-
-
-class LinearEquation(Equation):
-    def __init__(self, X, B, *, A=None, adj_type=None):
-        if isinstance(X, RHS) \
-                or (isinstance(X, Sequence) and len(X) > 0
-                    and isinstance(X[0], RHS)):
-            warnings.warn("LinearEquation(B, X, *, A=None, adj_type=None) "
-                          "signature is deprecated -- use "
-                          "LinearEquation(X, B, *, A=None, adj_type=None) "
-                          "instead",
-                          DeprecationWarning, stacklevel=2)
-            X, B = B, X
-
-        if is_function(X):
-            X = (X,)
-        if isinstance(B, RHS):
-            B = (B,)
-        if adj_type is None:
-            if A is None:
-                adj_type = "conjugate_dual"
-            else:
-                adj_type = "primal"
-
-        deps = []
-        dep_ids = {}
-        nl_deps = []
-        nl_dep_ids = {}
-
-        x_ids = set()
-        for x in X:
-            x_id = function_id(x)
-            if x_id in x_ids:
-                raise ValueError("Duplicate solve")
-            x_ids.add(x_id)
-            deps.append(x)
-            dep_ids[x_id] = len(deps) - 1
-
-        b_dep_indices = tuple([] for b in B)
-        b_nl_dep_indices = tuple([] for b in B)
-
-        for i, b in enumerate(B):
-            for dep in b.dependencies():
-                dep_id = function_id(dep)
-                if dep_id in x_ids:
-                    raise ValueError("Invalid dependency in linear Equation")
-                if dep_id not in dep_ids:
-                    deps.append(dep)
-                    dep_ids[dep_id] = len(deps) - 1
-                b_dep_indices[i].append(dep_ids[dep_id])
-            for dep in b.nonlinear_dependencies():
-                dep_id = function_id(dep)
-                if dep_id not in nl_dep_ids:
-                    nl_deps.append(dep)
-                    nl_dep_ids[dep_id] = len(nl_deps) - 1
-                b_nl_dep_indices[i].append(nl_dep_ids[dep_id])
-
-        b_dep_ids = tuple({function_id(b_dep): i
-                           for i, b_dep in enumerate(b.dependencies())}
-                          for b in B)
-
-        if A is not None:
-            A_dep_indices = []
-            A_nl_dep_indices = []
-            for dep in A.nonlinear_dependencies():
-                dep_id = function_id(dep)
-                if dep_id not in dep_ids:
-                    deps.append(dep)
-                    dep_ids[dep_id] = len(deps) - 1
-                A_dep_indices.append(dep_ids[dep_id])
-                if dep_id not in nl_dep_ids:
-                    nl_deps.append(dep)
-                    nl_dep_ids[dep_id] = len(nl_deps) - 1
-                A_nl_dep_indices.append(nl_dep_ids[dep_id])
-
-            A_nl_dep_ids = {function_id(A_nl_dep): i
-                            for i, A_nl_dep
-                            in enumerate(A.nonlinear_dependencies())}
-
-            if len(A.nonlinear_dependencies()) > 0:
-                A_x_indices = []
-                for x in X:
-                    x_id = function_id(x)
-                    if x_id not in nl_dep_ids:
-                        nl_deps.append(x)
-                        nl_dep_ids[x_id] = len(nl_deps) - 1
-                    A_x_indices.append(nl_dep_ids[x_id])
-
-        del x_ids, dep_ids, nl_dep_ids
-
-        super().__init__(
-            X, deps, nl_deps=nl_deps,
-            ic=A is not None and A.has_initial_condition(),
-            adj_ic=A is not None and A.adjoint_has_initial_condition(),
-            adj_type=adj_type)
-        self._B = tuple(B)
-        self._b_dep_indices = b_dep_indices
-        self._b_nl_dep_indices = b_nl_dep_indices
-        self._b_dep_ids = b_dep_ids
-        self._A = A
-        if A is not None:
-            self._A_dep_indices = A_dep_indices
-            self._A_nl_dep_indices = A_nl_dep_indices
-            self._A_nl_dep_ids = A_nl_dep_ids
-            if len(A.nonlinear_dependencies()) > 0:
-                self._A_x_indices = A_x_indices
-
-        self.add_referrer(*B)
-        if A is not None:
-            self.add_referrer(A)
-
-    def drop_references(self):
-        super().drop_references()
-        self._B = tuple(WeakAlias(b) for b in self._B)
-        if self._A is not None:
-            self._A = WeakAlias(self._A)
-
-    def forward_solve(self, X, deps=None):
-        if is_function(X):
-            X = (X,)
-        if deps is None:
-            deps = self.dependencies()
-
-        if self._A is None:
-            for x in X:
-                function_zero(x)
-            B = X
-        else:
-            def b_space_type(m):
-                space_type = function_space_type(
-                    self.X(m), rel_space_type=self.adj_X_type(m))
-                return conjugate_dual_space_type(space_type)
-
-            B = tuple(space_new(function_space(x), space_type=b_space_type(m))
-                      for m, x in enumerate(X))
-
-        for i, b in enumerate(self._B):
-            b.add_forward(B[0] if len(B) == 1 else B,
-                          [deps[j] for j in self._b_dep_indices[i]])
-
-        if self._A is not None:
-            self._A.forward_solve(X[0] if len(X) == 1 else X,
-                                  [deps[j] for j in self._A_dep_indices],
-                                  B[0] if len(B) == 1 else B)
-
-    _reset_adjoint_warning = False
-
-    def reset_adjoint(self):
-        if self._A is not None:
-            self._A.reset_adjoint()
-
-    _initialize_adjoint_warning = False
-
-    def initialize_adjoint(self, J, nl_deps):
-        if self._A is not None:
-            self._A.initialize_adjoint(J, nl_deps)
-
-    _finalize_adjoint_warning = False
-
-    def finalize_adjoint(self, J):
-        if self._A is not None:
-            self._A.finalize_adjoint(J)
-
-    def adjoint_jacobian_solve(self, adj_X, nl_deps, B):
-        if self._A is None:
-            return B
-        else:
-            return self._A.adjoint_solve(
-                adj_X, [nl_deps[j] for j in self._A_nl_dep_indices], B)
-
-    def adjoint_derivative_action(self, nl_deps, dep_index, adj_X):
-        if is_function(adj_X):
-            adj_X = (adj_X,)
-
-        eq_deps = self.dependencies()
-        if dep_index < 0 or dep_index >= len(eq_deps):
-            raise IndexError("dep_index out of bounds")
-        elif dep_index < len(self.X()):
-            if self._A is None:
-                return adj_X[dep_index]
-            else:
-                dep = eq_deps[dep_index]
-                F = function_new_conjugate_dual(dep)
-                self._A.adjoint_action([nl_deps[j]
-                                        for j in self._A_nl_dep_indices],
-                                       adj_X[0] if len(adj_X) == 1 else adj_X,
-                                       F, b_index=dep_index, method="assign")
-                return F
-        else:
-            dep = eq_deps[dep_index]
-            dep_id = function_id(dep)
-            F = function_new_conjugate_dual(dep)
-            assert len(self._B) == len(self._b_dep_ids)
-            for i, (b, b_dep_ids) in enumerate(zip(self._B, self._b_dep_ids)):
-                if dep_id in b_dep_ids:
-                    b_dep_index = b_dep_ids[dep_id]
-                else:
-                    continue
-                b_nl_deps = [nl_deps[j] for j in self._b_nl_dep_indices[i]]
-                b.subtract_adjoint_derivative_action(
-                    b_nl_deps, b_dep_index,
-                    adj_X[0] if len(adj_X) == 1 else adj_X,
-                    F)
-            if self._A is not None and dep_id in self._A_nl_dep_ids:
-                A_nl_dep_index = self._A_nl_dep_ids[dep_id]
-                A_nl_deps = [nl_deps[j] for j in self._A_nl_dep_indices]
-                X = [nl_deps[j] for j in self._A_x_indices]
-                self._A.adjoint_derivative_action(
-                    A_nl_deps, A_nl_dep_index,
-                    X[0] if len(X) == 1 else X,
-                    adj_X[0] if len(adj_X) == 1 else adj_X,
-                    F, method="add")
-            return F
-
-    def tangent_linear(self, M, dM, tlm_map):
-        X = self.X()
-
-        if self._A is None:
-            tlm_B = []
-        else:
-            tlm_B = self._A.tangent_linear_rhs(M, dM, tlm_map,
-                                               X[0] if len(X) == 1 else X)
-            if tlm_B is None:
-                tlm_B = []
-            elif isinstance(tlm_B, RHS):
-                tlm_B = [tlm_B]
-        for b in self._B:
-            tlm_b = b.tangent_linear_rhs(M, dM, tlm_map)
-            if tlm_b is None:
-                pass
-            elif isinstance(tlm_b, RHS):
-                tlm_B.append(tlm_b)
-            else:
-                tlm_B.extend(tlm_b)
-
-        if len(tlm_B) == 0:
-            return ZeroAssignment([tlm_map[x] for x in self.X()])
-        else:
-            return LinearEquation([tlm_map[x] for x in self.X()], tlm_B,
-                                  A=self._A, adj_type=self.adj_X_type())
-
-
-class Matrix(Referrer):
-    def __init__(self, nl_deps=None, *, has_ic_dep=None, ic=None, adj_ic=True):
-        if nl_deps is None:
-            nl_deps = []
-        if len({function_id(dep) for dep in nl_deps}) != len(nl_deps):
-            raise ValueError("Duplicate non-linear dependency")
-
-        if has_ic_dep is not None:
-            warnings.warn("has_ic_dep argument is deprecated -- use ic "
-                          "instead",
-                          DeprecationWarning, stacklevel=2)
-            if ic is not None:
-                raise TypeError("Cannot pass both has_ic_dep and ic arguments")
-            ic = has_ic_dep
-        elif ic is None:
-            ic = True
-
-        super().__init__()
-        self._nl_deps = tuple(nl_deps)
-        self._ic = ic
-        self._adj_ic = adj_ic
-
-    _reset_adjoint_warning = True
-    _initialize_adjoint_warning = True
-    _finalize_adjoint_warning = True
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-
-        if hasattr(cls, "reset_adjoint"):
-            if cls._reset_adjoint_warning:
-                warnings.warn("Matrix.reset_adjoint method is deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._reset_adjoint_warning = False
-            cls.reset_adjoint = lambda self: None
-
-        if hasattr(cls, "initialize_adjoint"):
-            if cls._initialize_adjoint_warning:
-                warnings.warn("Matrix.initialize_adjoint method is "
-                              "deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._initialize_adjoint_warning = False
-            cls.initialize_adjoint = lambda self, J, nl_deps: None
-
-        if hasattr(cls, "finalize_adjoint"):
-            if cls._finalize_adjoint_warning:
-                warnings.warn("Matrix.finalize_adjoint method is deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._finalize_adjoint_warning = False
-            cls.finalize_adjoint = lambda self, J: None
-
-        adj_solve_sig = inspect.signature(cls.adjoint_solve)
-        if tuple(adj_solve_sig.parameters.keys()) in [("self", "nl_deps", "b"),
-                                                      ("self", "nl_deps", "B")]:  # noqa: E501
-            warnings.warn("Matrix.adjoint_solve(self, nl_deps, b/B) method "
-                          "signature is deprecated",
-                          DeprecationWarning, stacklevel=2)
-
-            def adjoint_solve(self, adj_X, nl_deps, B):
-                return adjoint_solve_orig(self, nl_deps, B)
-            adjoint_solve_orig = cls.adjoint_solve
-            cls.adjoint_solve = adjoint_solve
-
-    def drop_references(self):
-        self._nl_deps = tuple(function_replacement(dep)
-                              for dep in self._nl_deps)
-
-    def nonlinear_dependencies(self):
-        return self._nl_deps
-
-    def has_initial_condition_dependency(self):
-        warnings.warn("Matrix.has_initial_condition_dependency method is "
-                      "deprecated -- use Matrix.has_initial_condition instead",
-                      DeprecationWarning, stacklevel=2)
-        return self._ic
-
-    def has_initial_condition(self):
-        return self._ic
-
-    def adjoint_has_initial_condition(self):
-        return self._adj_ic
-
-    def forward_action(self, nl_deps, X, B, method="assign"):
-        """
-        Evaluate the (forward) action of the matrix.
-
-        The form:
-            forward_action(self, nl_deps, x, b, method="assign")
-        should be used for matrices which act on a single function.
-
-        Arguments:
-
-        nl_deps      A sequence of functions defining the values of non-linear
-                     dependencies.
-        x/X          The argument of the matrix action.
-        b/B          The result of the matrix action.
-        method       (Optional) One of {"assign", "add", "sub"}.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def adjoint_action(self, nl_deps, adj_X, b, b_index=0, method="assign"):
-        """
-        Evaluate the adjoint action of the matrix.
-
-        The form:
-            adjoint_action(self, nl_deps, adj_x, b, b_index=0, method="assign")
-        should be used for matrices which act on a single function.
-
-        Arguments:
-
-        nl_deps      A sequence of functions defining the values of non-linear
-                     dependencies.
-        adj_x/adj_X  The argument of the matrix action.
-        b            The result of the matrix action.
-        b_index      (Optional) The element of the matrix action B to compute.
-        method       (Optional) One of {"assign", "add", "sub"}.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def forward_solve(self, X, nl_deps, B):
-        raise NotImplementedError("Method not overridden")
-
-    def adjoint_derivative_action(self, nl_deps, nl_dep_index, X, adj_X, b,
-                                  method="assign"):
-        """
-        Evaluate the action of the adjoint of a derivative of the matrix
-        action.
-
-        The form:
-            adjoint_derivative_action(self, nl_deps, nl_dep_index, x, adj_x, b,
-                                      method="assign")
-        should be used for matrices which act on a single function.
-
-        Arguments:
-
-        nl_deps      A sequence of functions defining the values of non-linear
-                     dependencies.
-        nl_dep_index The index of the dependency in
-                     self.nonlinear_dependencies() with respect to which a
-                     derivative should be taken.
-        x/X          The argument of the forward matrix action.
-        adj_x/adj_X  A function or sequence of functions on which the adjoint
-                     of the derivative acts.
-        b            The result.
-        method       (Optional) One of {"assign", "add", "sub"}.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def adjoint_solve(self, adj_X, nl_deps, B):
-        raise NotImplementedError("Method not overridden")
-
-    def tangent_linear_rhs(self, M, dM, tlm_map, X):
-        raise NotImplementedError("Method not overridden")
-
-
-class RHS(Referrer):
-    def __init__(self, deps, nl_deps=None):
-        dep_ids = set(map(function_id, deps))
-        if len(dep_ids) != len(deps):
-            raise ValueError("Duplicate dependency")
-
-        if nl_deps is None:
-            nl_deps = tuple(deps)
-        nl_dep_ids = set(map(function_id, nl_deps))
-        if len(nl_dep_ids) != len(nl_deps):
-            raise ValueError("Duplicate non-linear dependency")
-        if len(dep_ids.intersection(nl_dep_ids)) != len(nl_deps):
-            raise ValueError("Non-linear dependency is not a dependency")
-
-        super().__init__()
-        self._deps = tuple(deps)
-        self._nl_deps = tuple(nl_deps)
-
-    def drop_references(self):
-        self._deps = tuple(function_replacement(dep) for dep in self._deps)
-        self._nl_deps = tuple(function_replacement(dep)
-                              for dep in self._nl_deps)
-
-    def dependencies(self):
-        return self._deps
-
-    def nonlinear_dependencies(self):
-        return self._nl_deps
-
-    def add_forward(self, B, deps):
-        raise NotImplementedError("Method not overridden")
-
-    def subtract_adjoint_derivative_action(self, nl_deps, dep_index, adj_X, b):
-        raise NotImplementedError("Method not overridden")
-
-    def tangent_linear_rhs(self, M, dM, tlm_map):
-        raise NotImplementedError("Method not overridden")
 
 
 class MatrixActionSolver(LinearEquation):
