+++ tlm_adjoint/firedrake/firedrake_equations.py
@@ -27,8 +27,7 @@ from .backend_code_generator_interface import assemble, complex_mode, \
     matrix_multiply
 
 from ..caches import Cache
-from ..equation import Equation
-from ..equations import ZeroAssignment
+from ..equation import Equation, ZeroAssignment
 from ..tangent_linear import get_tangent_linear
 
 from .caches import form_dependencies, form_key, parameters_key
